requirements.txt
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3